Class do nothing constructor.std::stoul conversion

+++ src/UI_Helper.cpp
@@ -25,7 +25,7 @@ void UI_Helper<T>::search_vector(std::vector<std::wstring>* vec, std::wstring se
 
 template<class T>
 unsigned int UI_Helper<T>::get_id_from_wstring(std::wstring menu_entry){
-    unsigned int id = std::stoi(menu_entry.substr(0,10));
+	unsigned int id = std::stoul(menu_entry.substr(0, 10));
     return id;
 }
